Get carrion consumption rates

Returns a named vector with one entry for each species giving the rate in grams/year at which that species consumes carrion, plus a decompose entry for bacterial decomposition.

Usage

getCarrionConsumption(params, ...)

Arguments

params

A MizerParams object.

...

Unused

Value

A named vector with the consumption rates from all species and decomposition.

See also