mizerShelf: A mizer extension package for including detritus and carrion components.

This is an extension package for the mizer package (https://sizespectrum.org/mizer/) that makes it easy to set up a mizer model with a detritus and a carrion component, thereby allowing for more realistic modelling of the benthic component of fisheries on the continental shelf.

This package was developed to support the creation and exploration of a model for the fishery off the coast in the Northwest Mediterranean near Blanes which forms the basis of the paper "A model of size-spectrum dynamics to estimate the effects of improving fisheries selectivity and reducing discards in Mediterranean mixed demersal fisheries".

See also

Author

Maintainer: Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Authors:

Gustav Delius gustav.delius@york.ac.uk (ORCID) [copyright holder]

Silvia de Juan sdejuan@icm.csic.es (ORCID)

Francesc Maynou maynouf@icm.csic.es (ORCID)